added intelligence, aggressiveness and merged with visualisation locally
