+++ MAINCODE/main.py
@@ -1,25 +1,24 @@
 import random
 import numpy as np
 import csv
-import os
 import matplotlib.pyplot as plt
 import time
 from matplotlib.pyplot import imshow
+import os
 
 # Konstanten
 ENERGYCOSTS_MOVEMENT = 1
 ENERGYCOSTS_REPRODUCTION = 5
-START_ENERGY = 20
-WIDTH = 100
-HEIGHT = 100
-NUMBER_AGENTS = 50
-ROUNDS = 100
-FOOD_PERCENTAGE_BEGINNING = 1
-ADDITIONAL_FOOD_PERCENTAGE = 0.5
+START_ENERGY = 10
+WIDTH = 10
+HEIGHT = 10
+NUMBER_AGENTS = 10
+ROUNDS = 2
+FOOD_PERCENTAGE_BEGINNING = 0.1
+ADDITIONAL_FOOD_PERCENTAGE = 0
 SICKNESS_DURATION = ROUNDS // 10
 
-
-# Globaler Counter
+# Globaler Counter für die Nummerierung der Lebewesen
 agents_counter = NUMBER_AGENTS
 
 FOOD = {
@@ -121,28 +120,17 @@ class Agent:
                         new_y = max(0, min(HEIGHT - 1, self.position[1] + dy))
                         self.position = (new_x, new_y)
                        
-                    else:
-                        self.energy -= ENERGYCOSTS_MOVEMENT
-                        new_position = self.search_food(board)
-                        if new_position:
-                            self.position = new_position
-                        else:
-                            # Zufällige Bewegung: -1 oder 1, multipliziert mit der Kondition
-                            dx = random.choice([-1, 1]) * self.genetic['Kondition']
-                            dy = random.choice([-1, 1]) * self.genetic['Kondition']
-                            new_x = max(0, min(WIDTH - 1, self.position[0] + dx))
-                            new_y = max(0, min(HEIGHT - 1, self.position[1] + dy))
-                            self.position = (new_x, new_y)
-                self.covered_distance += 1
+                else:
+                    # Zufällige Bewegung: -1 oder 1, multipliziert mit der Kondition
+                    dx = random.choice([-1, 1]) * self.genetic['Kondition']
+                    dy = random.choice([-1, 1]) * self.genetic['Kondition']
+
+                    new_x = max(0, min(WIDTH - 1, self.position[0] + dx))
+                    new_y = max(0, min(HEIGHT - 1, self.position[1] + dy))
+                    self.position = (new_x, new_y)
+                    self.covered_distance += 1
             else:
                 return "deceased"
-
-    def check_for_sickness(self):
-        if self.sick:
-            self.sickness_duration -= 1
-            if self.sickness_duration <= 0:
-                self.sick = False
-                self.genetic["Kondition"] = self.previous_kondition
 
     def search_food(self, board):
         visibilityrange = self.genetic["Visibilityrange"]
@@ -150,9 +138,10 @@ class Agent:
             for dy in range(-visibilityrange, visibilityrange + 1):
                 x, y = self.position[0] + dx, self.position[1] + dy
 
+
                 if 0 <= x < WIDTH and 0 <= y < HEIGHT and board.food[x][y] is not None:
                     food_dict = board.food[x][y]
-                    # Prüfen, ob sich aggressive Agents in der Nähe befinden
+                    # Prüfen, ob     sich aggressive Agents in der Nähe befinden
                     aggressive_agents_nearby = self.check_for_aggressive_agents(board, x, y)
 
                     if aggressive_agents_nearby:
@@ -171,6 +160,7 @@ class Agent:
                         self.consuming_food(food_dict)  # Aufrufen von consuming_food mit dem Schlüssel
                         board.food[x][y] = None  # Entfernen der Nahrung von den Koordinaten
                         return (x, y)
+
         return None
 
     def check_for_aggressive_agents(self, board, x, y):
@@ -222,7 +212,7 @@ class Board:
         self.width = width
         self.height = height
         self.agents_list = []
-        self.food = np.full((width, height), None)
+        self.food = np.zeros((width, height))
         self.world = np.zeros((width, height))
 
     def add_agent(self, agents_to_add):
@@ -232,12 +222,26 @@ class Board:
         amount_fields = int(self.width * self.height * prozent)
         for _ in range(amount_fields):
             x, y = random.randint(0, self.width - 1), random.randint(0, self.height - 1)
-            food_type = random.choice(FOOD_KEYS)
-            food_dict = FOOD[food_type]
+            food_key = random.choice(FOOD_KEYS)
+            #food_dict = FOOD[food_type]
             # Speichern des Nahrungstyps und der zugehörigen Werte direkt im food-Array
-            self.food[x][y] = food_dict
+            self.food[x][y] = food_key
     
         
+    def remove_agents(self, agent):
+        #removing the agents in the list 'lebewesen'
+        self.agents_list.remove(agent)
+            
+
+    def place_agents(self):
+        #deleting the array so deceased agents will not be shown
+        self.world = np.zeros_like(self.world)
+        
+        #placing agents in array
+        for agent in self.agents_list:
+            x, y = agent.position
+            self.world[x][y] += 1
+
     def remove_agents(self, agent):
         #removing the agents in the list 'lebewesen'
         self.agents_list.remove(agent)
@@ -249,12 +253,20 @@ class Game:
         for i in range(1, NUMBER_AGENTS + 1):
             self.board.add_agent(Agent(i))
         self.board.place_food(FOOD_PERCENTAGE_BEGINNING)
-        
+    
         
             
     def run(self):
+        print("----------Round 0------------")
+        #fillng the world with agents for the start
+        self.board.place_agents()
+        
+        #visualizing the board
+        ###self.visualize_board(FOOD) 
         
         for round in range(ROUNDS):
+            #counter for Rounds
+            print(f"------------Round {round+1}------------")  
             
             #ending the simulation in case there are no agents left
             if len(self.board.agents_list) == 0:
@@ -262,19 +274,19 @@ class Game:
                 print("\nall agents deceased\n")
                 print("--------------------------")
                 
-                break
-            if round % 10 == 0:
-                self.board.place_food(ADDITIONAL_FOOD_PERCENTAGE)
-                print(self.board.food)
-                
+                break 
+            
+            
             for agent in self.board.agents_list[:]:
-                #bewegt die agents
-                result = agent.move(self.board)
 
+                #moves the agents
+                result = agent.move(self.board)
+                """for agent in self.board.agents_list:
+                    print(f"position: {agent.position}")"""
+                
                 #schaut ob der agent deceased ist, wenn ja, dann entfernt er diesen
                 if result == "deceased":
                     self.board.remove_agents(agent)
-
 
                 #wenn agent noch lebt wird die fortpflanzung weitergeführt
                     #vergleicht Agent mit potentiellen partnern
@@ -282,19 +294,36 @@ class Game:
                     for partner in self.board.agents_list:
                         if agent != partner:
                             agent.reproduce(partner, self.board)
+                
+                  
+            #placing additional food and all agents in every round
+            self.board.place_food(ADDITIONAL_FOOD_PERCENTAGE)
+            print(self.board.food)
+            self.board.place_agents()
+            #visualizing the board in every round
+            self.visualize_board(FOOD) 
+            
+            if round % 10 == 0:
+                #placing additional food and all agents every 10 rounds
+                self.board.place_food(ADDITIONAL_FOOD_PERCENTAGE)
+                self.board.place_agents()
+        
+                #visualizing the board every 10 rounds
+                self.visualize_board(FOOD)             
             
             
+        self.board.place_agents()  
+
         if self.saving is True:    ### Mögliches Laufzeitproblem: Doppelte For-Schleife sorgt für Quadratische Laufzeit O(n2)
             self.save_data()
 
         
     def save_data(self):
         # Generiere den Pfad für das "results" Verzeichnis im aktuellen Arbeitsverzeichnis
-        current_working_directory = os.getcwd() + "/MAINCODE/"
+        current_working_directory = os.getcwd()
         result_dir = os.path.join(current_working_directory, 'results')
         csv_index = 0
-        
-        # Erstelle das Verzeichnis, wenn es nicht existiert
+
         if not os.path.exists(result_dir):
             os.makedirs(result_dir)
 
@@ -303,7 +332,7 @@ class Game:
         while os.path.exists(filename):
             csv_index += 1
             filename = os.path.join(result_dir, f'agent_data_{csv_index}.csv')
-        
+
         # Schreibe die Agenten-Daten in die CSV-Datei
         with open(filename, 'w', newline='') as file:
             writer = csv.writer(file)
@@ -326,8 +355,72 @@ class Game:
                     agent.position,
                     agent.expelled
                 ])
+   
+    def visualize_board(self, food):
+        #5 seconds pause between each time visualizing 
+        #time.sleep(2)
+        
+        plt.rcParams["figure.figsize"] = [7.50, 3.50]
+        plt.rcParams["figure.autolayout"] = True
+        
+        dx, dy = 0.05, 0.05
+        x = np.arange(-3.0, 3.0, dx)
+        y = np.arange(-3.0, 3.0, dy)
+        extent = np.min(x), np.max(x), np.min(y), np.max(y)
+        
+        fig = plt.figure(frameon=False)
 
-                
+
+
+        data1 = self.board.food
+        plot1 = plt.imshow(data1, cmap="YlGn", interpolation='nearest', extent=extent)
+        
+        data2 = self.board.world
+        plot2 = plt.imshow(data2, cmap="YlOrRd",alpha = .7, interpolation='bilinear', extent=extent)
+        
+        
+        # set imshow outline to white
+        for spine in plot2.axes.spines.values():
+            spine.set_edgecolor("white")
+            
+        cb2 = plt.colorbar(plot2)
+        cb1 = plt.colorbar(plot1)
+        
+        # COLORBAR
+        # set colorbar label plus label color for agents
+        cb2.set_label('amount of agents', color="white")
+        cb1.set_label('amount of food per field', color="white")
+        # set colorbar tick color
+        cb2.ax.yaxis.set_tick_params(color="white")
+        cb1.ax.yaxis.set_tick_params(color="white")
+        
+        # set colorbar edgecolor 
+        cb2.outline.set_edgecolor("white")
+        cb1.outline.set_edgecolor("white")
+        
+        # set colorbar ticklabels
+        plt.setp(plt.getp(cb2.ax.axes, 'yticklabels'), color="white")
+        plt.setp(plt.getp(cb1.ax.axes, 'yticklabels'), color="white")
+
+        
+        plt.title("Distribution of food and agents in the world",color = "white")
+        plt.show()
+        
+        #showing food array to check visualization (optional)
+        #print(self.board.food)
+        #showing world array to check visualization (optional)
+        #print(self.board.world)
+        
+        #showing number of angents still living
+        print(f"number of agents: {len(self.board.agents_list)}")
+        
+        #checking if there is food left in the world
+        #if not agents will probably die in a couple of rounds
+        if self.board.food.any() >=1 :
+            print('still some food left')
+        else:
+            print("no food left")            
+    
     
 # Counter einfügen wie oft sich ein Agents fortgepflanzt hat 
 # Stammesangehörigkeit ausbessern: Aktuell Tupel für Stamm des Kindes
@@ -339,3 +432,4 @@ if __name__ == "__main__":
     game.run()
     script_time = np.round(time.time() - start, 2)
     print(script_time)
+
