--- MAINCODE/results/agent_data_0.csv
@@ -1,15 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Position
-1,2,3,1,1,6,"(5, 2)"
-2,1,2,3,1,9,"(9, 9)"
-3,2,1,3,2,9,"(8, 9)"
-4,2,3,1,2,3,"(9, 3)"
-5,2,1,3,1,9,"(9, 5)"
-6,2,2,2,0,9,"(7, 4)"
-7,3,1,3,1,9,"(9, 5)"
-8,1,1,1,0,3,"(1, 9)"
-9,3,1,3,1,9,"(8, 9)"
-10,3,3,1,0,8,"(3, 3)"
-12,"((2, 2), 2)",0,0,0,1,"(8, 1)"
-13,"(2, 1)",1,0,0,0,"(8, 2)"
-14,"(3, 2)",0,1,0,0,"(7, 6)"
-15,"(3, 2)",0,1,0,0,"(9, 7)"
