--- MAINCODE/results/agent_data_1.csv
@@ -1,11 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Position
-1,3,2,1,0,10,"(90, 23)"
-2,2,1,1,0,10,"(87, 86)"
-3,1,2,2,0,10,"(23, 0)"
-4,1,1,3,0,10,"(57, 0)"
-5,2,1,2,0,10,"(78, 47)"
-6,2,2,2,0,10,"(0, 76)"
-7,2,2,1,0,10,"(16, 79)"
-8,1,1,1,0,10,"(77, 22)"
-9,1,1,2,0,10,"(28, 10)"
-10,3,3,2,0,10,"(1, 0)"
