--- MAINCODE/results/agent_data_10.csv
@@ -1,51 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
-1,3,2,2,0,0,0,,,"(82, 7)"
-2,3,1,3,0,0,0,,,"(5, 59)"
-3,1,2,1,0,0,0,,,"(16, 89)"
-4,3,1,2,0,0,0,,,"(8, 19)"
-5,3,1,3,0,0,0,,,"(65, 56)"
-6,2,2,3,0,0,0,,,"(58, 75)"
-7,3,3,3,0,0,0,,,"(42, 22)"
-8,3,1,3,0,0,0,,,"(80, 46)"
-9,2,1,2,0,0,0,,,"(11, 10)"
-10,2,1,3,0,0,0,,,"(46, 13)"
-11,3,2,3,0,0,0,,,"(75, 10)"
-12,1,1,1,0,0,0,,,"(89, 38)"
-13,1,1,1,0,0,0,,,"(73, 35)"
-14,3,1,2,0,0,0,,,"(23, 68)"
-15,3,3,1,0,0,0,,,"(26, 46)"
-16,3,1,2,0,0,0,,,"(23, 34)"
-17,1,3,2,0,0,0,,,"(27, 27)"
-18,3,2,2,0,0,0,,,"(87, 8)"
-19,2,1,2,0,0,0,,,"(8, 70)"
-20,3,1,1,0,0,0,,,"(38, 80)"
-21,2,3,1,0,0,0,,,"(4, 22)"
-22,2,3,3,0,0,0,,,"(53, 79)"
-23,3,1,2,0,0,0,,,"(99, 41)"
-24,2,2,3,0,0,0,,,"(15, 74)"
-25,1,3,2,0,0,0,,,"(79, 17)"
-26,2,2,3,0,0,0,,,"(99, 9)"
-27,3,3,2,0,0,0,,,"(58, 92)"
-28,1,1,3,0,0,0,,,"(98, 30)"
-29,2,2,3,0,0,0,,,"(57, 29)"
-30,2,3,1,0,0,0,,,"(75, 72)"
-31,2,1,2,0,0,0,,,"(54, 18)"
-32,3,2,2,0,0,0,,,"(13, 99)"
-33,2,1,1,0,0,0,,,"(12, 39)"
-34,3,1,3,0,0,0,,,"(69, 37)"
-35,2,3,1,0,0,0,,,"(39, 43)"
-36,2,1,3,0,0,0,,,"(51, 81)"
-37,2,1,1,0,0,0,,,"(79, 9)"
-38,1,1,1,0,0,0,,,"(25, 56)"
-39,2,2,1,0,0,0,,,"(13, 65)"
-40,2,2,2,0,0,0,,,"(95, 10)"
-41,3,3,3,0,0,0,,,"(63, 79)"
-42,1,3,1,0,0,0,,,"(49, 22)"
-43,3,2,2,0,0,0,,,"(5, 40)"
-44,3,1,2,0,0,0,,,"(30, 84)"
-45,2,1,2,0,0,0,,,"(71, 98)"
-46,2,1,1,0,0,0,,,"(74, 70)"
-47,2,2,2,0,0,0,,,"(51, 5)"
-48,1,3,2,0,0,0,,,"(99, 36)"
-49,1,1,2,0,0,0,,,"(38, 2)"
-50,1,1,1,0,0,0,,,"(18, 52)"
