--- MAINCODE/results/agent_data_11.csv
@@ -1,57 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
-1,1,3,3,0,0,0,,,"(40, 79)"
-2,1,2,3,0,0,0,,,"(92, 9)"
-3,3,2,3,0,0,0,,,"(36, 52)"
-4,3,2,1,0,0,0,,,"(69, 83)"
-5,3,3,1,0,0,0,,,"(60, 28)"
-6,3,3,2,0,0,0,,,"(30, 60)"
-7,2,1,2,0,0,0,,,"(19, 10)"
-8,3,3,2,0,0,0,,,"(21, 60)"
-9,3,1,3,0,0,0,,,"(55, 19)"
-10,1,1,1,0,0,0,,,"(58, 74)"
-11,2,2,3,0,0,0,,,"(37, 89)"
-12,1,3,1,0,0,0,,,"(3, 15)"
-13,1,2,1,0,0,0,,,"(68, 31)"
-14,2,1,3,0,0,0,,,"(6, 22)"
-15,1,1,2,0,0,0,,,"(44, 23)"
-16,3,1,3,0,0,0,,,"(69, 67)"
-17,2,1,1,0,0,0,,,"(13, 22)"
-18,1,1,2,0,0,0,,,"(30, 19)"
-19,3,1,3,0,0,0,,,"(11, 90)"
-20,2,1,2,0,0,0,,,"(9, 59)"
-21,1,3,1,0,0,0,,,"(17, 39)"
-22,3,2,3,0,0,0,,,"(11, 16)"
-23,1,1,2,0,0,0,,,"(93, 88)"
-24,3,3,2,0,0,0,,,"(90, 65)"
-25,1,1,1,0,0,0,,,"(46, 75)"
-26,1,1,3,0,0,0,,,"(95, 43)"
-27,3,3,2,0,0,0,,,"(85, 9)"
-28,3,3,3,0,0,0,,,"(80, 31)"
-29,2,1,1,0,0,0,,,"(72, 41)"
-30,2,1,2,0,0,0,,,"(81, 67)"
-31,3,2,2,0,0,0,,,"(54, 2)"
-32,2,1,3,0,0,0,,,"(55, 9)"
-33,3,1,3,0,0,0,,,"(64, 21)"
-34,2,2,1,0,0,0,,,"(90, 68)"
-35,1,1,2,0,0,0,,,"(32, 64)"
-36,3,1,2,6,0,0,,,"(39, 51)"
-37,1,3,3,0,0,0,,,"(82, 27)"
-38,3,2,3,0,0,0,,,"(41, 35)"
-39,3,3,2,0,0,0,,,"(14, 66)"
-40,2,1,1,0,0,0,,,"(61, 89)"
-41,3,2,2,6,0,0,,,"(39, 51)"
-42,2,3,2,0,0,0,,,"(63, 2)"
-43,1,1,3,0,0,0,,,"(49, 43)"
-44,2,2,1,0,0,0,,,"(89, 86)"
-45,1,3,2,0,0,0,,,"(60, 18)"
-46,1,3,1,0,0,0,,,"(63, 26)"
-47,1,2,2,0,0,0,,,"(80, 33)"
-48,1,3,1,0,0,0,,,"(23, 98)"
-49,1,2,3,0,0,0,,,"(18, 89)"
-50,3,1,1,0,0,0,,,"(94, 18)"
-51,3,0,1,0,0,0,36,41,"(63, 32)"
-52,3,0,1,0,0,0,41,36,"(48, 10)"
-53,3,0,1,0,0,0,36,41,"(80, 81)"
-54,3,0,1,0,0,0,41,36,"(1, 33)"
-55,3,0,1,0,0,0,36,41,"(86, 68)"
-56,3,0,1,0,0,0,41,36,"(59, 8)"
