--- MAINCODE/results/agent_data_12.csv
@@ -1,51 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
-1,2,3,1,3,0,0,0,,,"(50, 8)"
-2,1,3,2,1,0,0,0,,,"(68, 86)"
-3,1,2,3,3,0,0,0,,,"(18, 58)"
-4,1,3,1,2,0,0,0,,,"(22, 20)"
-5,3,2,2,1,0,0,0,,,"(96, 85)"
-6,2,1,1,1,0,0,0,,,"(97, 98)"
-7,2,1,3,3,0,0,0,,,"(89, 31)"
-8,2,3,2,1,0,0,0,,,"(27, 65)"
-9,3,1,2,2,0,0,0,,,"(8, 62)"
-10,3,1,1,3,0,0,0,,,"(70, 6)"
-11,3,2,1,2,0,0,0,,,"(34, 47)"
-12,3,2,2,3,0,0,0,,,"(60, 62)"
-13,3,1,2,3,0,0,0,,,"(39, 23)"
-14,3,1,2,3,0,0,0,,,"(20, 33)"
-15,1,1,1,1,0,0,0,,,"(63, 48)"
-16,1,2,3,1,0,0,0,,,"(14, 92)"
-17,3,3,1,2,0,0,0,,,"(29, 67)"
-18,3,2,3,2,0,0,0,,,"(41, 22)"
-19,2,2,3,3,0,0,0,,,"(27, 76)"
-20,1,1,3,1,0,0,0,,,"(84, 31)"
-21,3,1,1,1,0,0,0,,,"(30, 66)"
-22,3,3,3,3,0,0,0,,,"(56, 11)"
-23,3,3,2,3,0,0,0,,,"(13, 54)"
-24,3,2,3,3,0,0,0,,,"(23, 31)"
-25,2,3,3,1,0,0,0,,,"(3, 30)"
-26,3,1,1,3,0,0,0,,,"(74, 3)"
-27,3,3,1,1,0,0,0,,,"(62, 38)"
-28,2,3,3,1,0,0,0,,,"(11, 53)"
-29,1,1,2,3,0,0,0,,,"(36, 23)"
-30,1,3,3,1,0,0,0,,,"(61, 17)"
-31,3,2,2,1,0,0,0,,,"(16, 15)"
-32,3,1,3,2,0,0,0,,,"(27, 4)"
-33,3,3,3,3,0,0,0,,,"(38, 48)"
-34,3,2,3,1,0,0,0,,,"(93, 43)"
-35,1,3,1,2,0,0,0,,,"(91, 71)"
-36,3,2,2,2,0,0,0,,,"(72, 63)"
-37,3,2,1,2,0,0,0,,,"(95, 84)"
-38,1,3,1,3,0,0,0,,,"(92, 48)"
-39,3,1,1,2,0,0,0,,,"(98, 87)"
-40,3,2,2,2,0,0,0,,,"(45, 98)"
-41,2,1,1,2,0,0,0,,,"(22, 29)"
-42,1,3,3,3,0,0,0,,,"(53, 40)"
-43,1,2,3,2,0,0,0,,,"(45, 76)"
-44,3,3,2,2,0,0,0,,,"(45, 43)"
-45,2,1,2,1,0,0,0,,,"(13, 10)"
-46,1,3,3,2,0,0,0,,,"(87, 76)"
-47,2,3,1,3,0,0,0,,,"(44, 38)"
-48,1,3,1,3,0,0,0,,,"(41, 76)"
-49,1,2,2,2,0,0,0,,,"(40, 91)"
-50,2,3,3,2,0,0,0,,,"(9, 80)"
