--- MAINCODE/results/agent_data_13.csv
@@ -1,51 +1,0 @@
-Number ,Tribe ,Condition ,Visibility Range ,Metabolism ,Reproduction Counter ,Consume Counter ,Sickness Counter ,Parent A ,Parent B ,Position 
-1,3,1,1,3,0,0,0,,,"(18, 78)"
-2,3,1,3,3,0,0,0,,,"(38, 57)"
-3,1,3,3,3,0,0,0,,,"(93, 67)"
-4,1,3,1,2,0,0,0,,,"(20, 97)"
-5,3,3,2,3,0,0,0,,,"(66, 41)"
-6,2,1,1,1,0,0,0,,,"(30, 19)"
-7,1,1,1,3,0,0,0,,,"(89, 14)"
-8,3,1,3,3,0,0,0,,,"(73, 2)"
-9,2,1,2,3,0,0,0,,,"(49, 49)"
-10,2,3,1,3,0,0,0,,,"(85, 77)"
-11,1,3,3,3,0,0,0,,,"(59, 49)"
-12,3,1,3,3,0,0,0,,,"(39, 25)"
-13,3,2,2,1,0,0,0,,,"(29, 43)"
-14,1,3,2,2,0,0,0,,,"(5, 72)"
-15,3,2,2,1,0,0,0,,,"(77, 39)"
-16,1,3,2,1,0,0,0,,,"(86, 87)"
-17,3,1,3,3,0,0,0,,,"(82, 31)"
-18,3,2,1,1,0,0,0,,,"(22, 23)"
-19,2,2,3,2,0,0,0,,,"(37, 99)"
-20,2,2,2,1,0,0,0,,,"(51, 16)"
-21,3,3,3,2,0,0,0,,,"(94, 74)"
-22,1,2,2,3,0,0,0,,,"(95, 15)"
-23,2,1,2,3,0,0,0,,,"(98, 31)"
-24,2,2,1,2,0,0,0,,,"(31, 91)"
-25,3,1,1,1,0,0,0,,,"(19, 26)"
-26,1,1,3,2,0,0,0,,,"(35, 18)"
-27,2,2,1,3,0,0,0,,,"(15, 88)"
-28,2,1,3,2,0,0,0,,,"(85, 98)"
-29,1,1,2,3,0,0,0,,,"(6, 74)"
-30,2,1,2,3,0,0,0,,,"(52, 83)"
-31,2,3,1,3,0,0,0,,,"(49, 40)"
-32,3,2,3,3,0,0,0,,,"(64, 2)"
-33,2,3,2,1,0,0,0,,,"(57, 69)"
-34,3,3,1,3,0,0,0,,,"(38, 26)"
-35,1,3,3,2,0,0,0,,,"(51, 62)"
-36,3,2,3,3,0,0,0,,,"(22, 31)"
-37,3,3,1,3,0,0,0,,,"(67, 36)"
-38,1,2,1,1,0,0,0,,,"(54, 20)"
-39,1,1,3,3,0,0,0,,,"(16, 39)"
-40,2,2,3,1,0,0,0,,,"(68, 69)"
-41,1,3,1,2,0,0,0,,,"(10, 59)"
-42,2,3,2,3,0,0,0,,,"(81, 33)"
-43,1,3,1,2,0,0,0,,,"(22, 82)"
-44,3,2,3,1,0,0,0,,,"(73, 36)"
-45,3,3,3,2,0,0,0,,,"(15, 72)"
-46,2,2,3,1,0,0,0,,,"(79, 81)"
-47,2,1,3,2,0,0,0,,,"(90, 59)"
-48,3,3,3,3,0,0,0,,,"(7, 33)"
-49,3,1,2,1,0,0,0,,,"(36, 50)"
-50,2,3,3,3,0,0,0,,,"(1, 96)"
