--- MAINCODE/results/agent_data_14.csv
@@ -1,51 +1,0 @@
-Number ,Tribe ,Condition ,Visibility Range ,Metabolism ,Covered Distance ,Reproduction Counter ,Consume Counter ,Sickness Counter ,Parent A ,Parent B ,Position 
-1,2,2,1,3,0,0,0,0,,,"(64, 43)"
-2,1,2,1,1,0,0,0,0,,,"(79, 54)"
-3,3,3,2,2,0,0,0,0,,,"(24, 91)"
-4,1,1,3,1,0,0,0,0,,,"(94, 77)"
-5,2,3,1,1,0,0,0,0,,,"(19, 31)"
-6,3,3,3,2,0,0,0,0,,,"(65, 71)"
-7,1,1,1,1,0,0,0,0,,,"(33, 29)"
-8,3,2,1,3,0,0,0,0,,,"(31, 31)"
-9,1,2,1,3,0,0,0,0,,,"(64, 40)"
-10,2,3,3,1,0,0,0,0,,,"(73, 82)"
-11,1,1,2,2,0,0,0,0,,,"(49, 75)"
-12,2,2,1,3,0,0,0,0,,,"(60, 60)"
-13,2,1,2,3,0,0,0,0,,,"(66, 62)"
-14,1,2,1,1,0,0,0,0,,,"(55, 12)"
-15,1,1,1,1,0,0,0,0,,,"(38, 66)"
-16,1,1,3,1,0,0,0,0,,,"(28, 14)"
-17,3,3,3,2,0,0,0,0,,,"(34, 54)"
-18,1,1,3,3,0,0,0,0,,,"(61, 90)"
-19,2,1,1,2,0,0,0,0,,,"(92, 25)"
-20,3,2,3,2,0,0,0,0,,,"(83, 83)"
-21,2,3,2,2,0,0,0,0,,,"(37, 20)"
-22,3,1,1,1,0,0,0,0,,,"(60, 29)"
-23,1,1,2,3,0,0,0,0,,,"(87, 27)"
-24,2,3,3,1,0,0,0,0,,,"(94, 53)"
-25,3,1,1,1,0,0,0,0,,,"(82, 78)"
-26,1,1,1,1,0,0,0,0,,,"(83, 39)"
-27,2,1,3,3,0,0,0,0,,,"(39, 17)"
-28,1,3,3,2,0,0,0,0,,,"(91, 60)"
-29,2,2,1,2,0,0,0,0,,,"(80, 35)"
-30,1,2,2,2,0,0,0,0,,,"(24, 83)"
-31,2,2,1,3,0,0,0,0,,,"(12, 66)"
-32,2,3,1,2,0,0,0,0,,,"(7, 47)"
-33,3,3,2,1,0,0,0,0,,,"(74, 66)"
-34,2,2,2,3,0,0,0,0,,,"(39, 8)"
-35,3,3,1,3,0,0,0,0,,,"(90, 70)"
-36,3,3,2,2,0,0,0,0,,,"(12, 29)"
-37,2,1,3,3,0,0,0,0,,,"(68, 80)"
-38,2,1,2,3,0,0,0,0,,,"(58, 74)"
-39,2,1,2,1,0,0,0,0,,,"(86, 18)"
-40,2,1,3,1,0,0,0,0,,,"(12, 46)"
-41,1,2,3,2,0,0,0,0,,,"(28, 12)"
-42,3,2,2,2,0,0,0,0,,,"(37, 47)"
-43,1,3,1,2,0,0,0,0,,,"(50, 80)"
-44,1,1,3,1,0,0,0,0,,,"(79, 92)"
-45,1,3,2,1,0,0,0,0,,,"(42, 11)"
-46,2,1,3,3,0,0,0,0,,,"(18, 2)"
-47,2,3,2,1,0,0,0,0,,,"(15, 29)"
-48,2,3,3,2,0,0,0,0,,,"(6, 55)"
-49,3,2,3,2,0,0,0,0,,,"(21, 90)"
-50,2,1,1,3,0,0,0,0,,,"(66, 21)"
