--- MAINCODE/results/agent_data_15.csv
@@ -1,69 +1,0 @@
-Number ,Tribe ,Condition ,Visibility Range ,Metabolism ,Covered Distance ,Reproduction Counter ,Consume Counter ,Sickness Counter ,Parent A ,Parent B ,Position 
-1,1,1,2,1,50,0,10,4,,,"(23, 58)"
-2,1,2,3,1,19,0,19,0,,,"(38, 1)"
-3,2,2,3,1,18,1,18,0,,,"(40, 0)"
-4,3,2,2,1,18,0,18,0,,,"(48, 0)"
-5,1,2,2,2,28,0,28,0,,,"(6, 22)"
-6,2,3,1,3,49,6,49,0,,,"(2, 17)"
-7,1,1,3,1,62,0,6,6,,,"(46, 76)"
-8,3,1,2,3,82,0,10,8,,,"(70, 57)"
-9,3,2,2,1,20,1,17,0,,,"(1, 12)"
-10,1,2,2,2,28,0,28,0,,,"(31, 0)"
-11,2,1,1,1,66,0,8,6,,,"(31, 34)"
-12,2,2,1,3,40,1,40,0,,,"(1, 44)"
-13,1,1,1,2,70,0,11,6,,,"(52, 36)"
-14,3,1,1,2,79,0,10,7,,,"(37, 0)"
-15,3,1,2,2,31,0,31,0,,,"(3, 0)"
-16,2,1,2,1,52,0,11,5,,,"(26, 62)"
-17,1,1,2,1,60,0,10,6,,,"(7, 0)"
-18,3,1,2,2,78,0,9,7,,,"(49, 76)"
-19,3,3,2,2,30,0,30,0,,,"(34, 22)"
-20,3,1,1,3,83,0,13,7,,,"(46, 51)"
-21,1,1,2,2,73,0,12,7,,,"(16, 0)"
-22,1,1,3,2,81,0,8,8,,,"(72, 73)"
-23,1,1,3,3,78,4,16,7,,,"(2, 2)"
-24,2,1,1,1,54,6,8,5,,,"(1, 15)"
-25,3,2,1,2,27,0,27,0,,,"(43, 58)"
-26,3,3,2,1,18,0,18,0,,,"(44, 60)"
-27,2,1,1,2,85,0,9,8,,,"(38, 29)"
-28,1,2,2,2,29,0,29,0,,,"(22, 12)"
-29,1,1,2,3,81,0,11,8,,,"(35, 63)"
-30,2,1,2,3,47,4,47,0,,,"(1, 8)"
-31,1,1,2,1,50,2,10,5,,,"(0, 7)"
-32,1,3,2,2,30,0,30,0,,,"(1, 6)"
-33,3,3,2,1,58,0,8,5,,,"(16, 33)"
-34,3,2,1,3,41,0,41,0,,,"(21, 54)"
-35,2,1,1,3,80,0,11,7,,,"(71, 74)"
-36,2,1,2,3,84,0,12,8,,,"(63, 0)"
-37,3,2,1,3,90,1,10,8,,,"(0, 66)"
-38,3,1,1,2,77,0,8,7,,,"(40, 88)"
-39,2,1,3,2,72,0,12,6,,,"(50, 41)"
-40,1,1,1,1,52,0,8,5,,,"(78, 49)"
-41,2,1,2,3,84,0,14,8,,,"(69, 20)"
-42,2,3,3,3,46,2,46,0,,,"(0, 21)"
-43,2,3,1,3,43,0,42,0,,,"(1, 40)"
-44,1,1,3,3,91,2,11,9,,,"(1, 43)"
-45,3,1,2,3,82,5,12,7,,,"(1, 2)"
-46,2,2,2,2,72,0,12,6,,,"(71, 0)"
-47,1,1,3,2,71,0,15,6,,,"(44, 1)"
-48,2,3,1,1,16,0,16,0,,,"(24, 69)"
-49,2,1,2,3,44,0,44,0,,,"(0, 22)"
-50,3,1,3,1,42,1,13,3,,,"(52, 1)"
-51,3,1,1,0,65,0,7,6,3,50,"(58, 6)"
-52,2,0,1,1,28,0,8,2,42,45,"(54, 51)"
-53,3,1,1,1,46,0,6,5,45,42,"(80, 36)"
-54,1,1,1,1,28,0,8,3,37,44,"(57, 91)"
-55,1,0,1,1,34,0,4,3,30,23,"(48, 32)"
-56,1,1,1,1,28,0,3,3,31,30,"(5, 1)"
-57,2,1,1,1,23,0,3,3,30,31,"(42, 13)"
-58,2,1,1,1,7,0,2,1,30,9,"(97, 39)"
-59,2,0,0,1,1,0,1,0,6,24,"(14, 47)"
-60,2,1,0,0,3,0,1,1,24,6,"(74, 56)"
-61,3,1,1,1,1,0,1,1,45,23,"(31, 42)"
-62,2,1,0,1,1,0,1,1,6,24,"(59, 16)"
-63,2,1,0,1,3,0,2,1,24,6,"(18, 84)"
-64,2,1,0,0,1,0,1,0,6,24,"(16, 9)"
-65,3,0,1,1,1,0,1,0,23,45,"(1, 63)"
-66,2,1,0,1,1,0,1,1,24,6,"(25, 51)"
-67,1,1,1,1,1,0,1,1,45,23,"(41, 38)"
-68,1,0,1,1,1,0,1,0,44,12,"(50, 5)"
