--- MAINCODE/results/agent_data_16.csv
@@ -1,100 +1,0 @@
-Number, Tribe, Condition, Visibility Range, Metabolism, Covered Distance, Reproduction Counter, Consume Counter, Sickness Counter, Parent A, Parent B, Position
-1,3,3,3,1,15,0,15,0,,,"(10, 14)"
-2,1,3,2,1,18,0,18,0,,,"(2, 54)"
-3,3,1,1,2,78,0,10,7,,,"(0, 87)"
-4,1,2,2,1,17,0,17,0,,,"(2, 12)"
-5,2,1,1,1,62,0,9,6,,,"(0, 61)"
-6,1,3,2,3,91,0,11,8,,,"(10, 1)"
-7,3,1,2,1,70,0,8,7,,,"(29, 2)"
-8,2,1,1,1,60,0,8,6,,,"(74, 15)"
-9,3,1,2,2,79,0,11,7,,,"(1, 30)"
-10,3,1,1,2,79,0,9,7,,,"(56, 41)"
-11,1,1,1,2,70,0,13,6,,,"(54, 11)"
-12,3,3,1,2,30,0,30,0,,,"(65, 0)"
-13,1,3,1,2,33,0,30,0,,,"(7, 3)"
-14,3,1,3,3,87,0,12,8,,,"(50, 45)"
-15,2,2,2,3,48,7,45,0,,,"(2, 10)"
-16,2,1,3,2,28,0,28,0,,,"(3, 10)"
-17,3,2,3,1,17,0,17,0,,,"(30, 18)"
-18,1,3,2,3,43,4,43,0,,,"(4, 7)"
-19,2,1,2,3,79,0,19,7,,,"(11, 54)"
-20,2,1,2,3,82,0,13,7,,,"(61, 11)"
-21,2,1,2,3,44,3,40,0,,,"(0, 1)"
-22,2,2,3,3,44,0,44,0,,,"(3, 11)"
-23,1,1,2,1,60,15,10,6,,,"(55, 43)"
-24,3,1,3,2,79,0,9,8,,,"(31, 45)"
-25,2,1,3,1,56,0,9,5,,,"(2, 3)"
-26,1,1,2,1,62,0,8,6,,,"(19, 65)"
-27,2,1,3,1,20,8,20,0,,,"(5, 4)"
-28,1,1,3,1,58,0,8,6,,,"(47, 0)"
-29,3,3,3,3,46,0,46,0,,,"(1, 20)"
-30,3,1,2,3,83,0,16,7,,,"(0, 36)"
-31,3,1,2,3,85,0,12,8,,,"(71, 65)"
-32,2,3,2,1,16,14,16,0,,,"(26, 0)"
-33,2,3,3,2,28,0,28,0,,,"(2, 14)"
-34,1,3,1,3,46,0,46,0,,,"(6, 3)"
-35,3,1,3,1,59,0,9,6,,,"(0, 42)"
-36,2,2,3,2,29,0,29,0,,,"(0, 30)"
-37,1,2,2,2,74,8,14,6,,,"(0, 49)"
-38,1,2,2,2,74,3,14,6,,,"(0, 50)"
-39,2,3,3,3,81,14,11,7,,,"(25, 0)"
-40,3,1,3,3,83,0,13,8,,,"(15, 0)"
-41,1,1,1,2,66,0,11,6,,,"(66, 89)"
-42,3,1,1,2,68,0,15,6,,,"(59, 73)"
-43,2,1,1,1,68,0,8,7,,,"(64, 68)"
-44,1,1,1,1,58,15,8,6,,,"(64, 51)"
-45,2,1,1,2,76,0,10,7,,,"(80, 72)"
-46,1,1,1,2,72,0,12,7,,,"(13, 11)"
-47,2,2,3,3,42,0,42,0,,,"(1, 23)"
-48,1,2,1,1,51,0,11,4,,,"(34, 27)"
-49,2,1,2,3,81,0,15,7,,,"(39, 19)"
-50,2,1,1,3,83,0,13,8,,,"(48, 46)"
-51,1,0,0,0,51,0,4,4,23,44,"(6, 61)"
-53,1,0,0,0,44,0,4,2,23,44,"(75, 13)"
-54,2,1,1,0,40,0,4,4,32,39,"(59, 38)"
-55,2,1,1,0,28,0,8,2,39,32,"(48, 10)"
-56,1,0,0,0,42,0,4,0,44,23,"(41, 47)"
-57,1,0,0,0,49,0,4,1,23,44,"(23, 77)"
-58,2,1,1,0,43,0,6,4,32,39,"(5, 66)"
-59,2,1,1,0,38,0,8,4,39,32,"(39, 15)"
-60,1,0,0,0,47,0,4,1,44,23,"(4, 83)"
-61,1,0,0,0,54,0,2,2,23,44,"(45, 47)"
-62,2,1,1,1,38,0,6,4,32,39,"(26, 53)"
-63,2,0,1,0,36,0,6,3,39,32,"(22, 61)"
-64,1,1,0,0,42,0,4,3,44,23,"(55, 26)"
-65,1,0,0,0,43,0,4,1,23,44,"(28, 19)"
-66,2,0,1,1,36,0,6,3,32,39,"(83, 63)"
-67,2,1,1,1,26,0,6,3,39,32,"(68, 24)"
-68,1,0,0,0,47,0,3,1,44,23,"(31, 39)"
-69,1,0,0,0,50,0,2,1,23,44,"(96, 72)"
-70,2,1,1,1,36,0,6,4,32,39,"(51, 9)"
-71,2,0,1,0,36,0,6,3,39,32,"(28, 1)"
-72,1,1,0,0,36,0,4,3,44,23,"(53, 92)"
-73,1,0,0,0,47,0,2,2,23,44,"(31, 46)"
-74,2,1,1,0,28,0,8,2,32,39,"(0, 75)"
-75,2,1,1,0,41,0,4,4,39,32,"(73, 82)"
-76,1,0,0,0,39,0,4,1,44,23,"(16, 75)"
-77,1,0,0,0,48,0,3,1,23,44,"(26, 29)"
-78,2,0,1,0,44,0,4,4,32,39,"(4, 60)"
-79,2,1,1,0,36,5,6,4,39,32,"(0, 48)"
-80,1,1,1,1,2,0,2,1,38,37,"(46, 4)"
-81,1,1,1,1,9,0,1,1,37,38,"(4, 49)"
-82,1,1,1,1,3,0,2,1,38,37,"(84, 42)"
-83,1,1,0,0,4,0,2,1,37,79,"(0, 99)"
-84,1,0,0,0,6,0,1,0,79,37,"(50, 50)"
-85,1,0,0,0,5,0,1,0,37,79,"(23, 25)"
-86,2,1,1,1,6,0,1,1,21,27,"(46, 16)"
-87,2,1,1,1,8,0,1,1,27,21,"(95, 72)"
-88,2,1,1,1,2,0,2,1,15,21,"(30, 86)"
-89,2,1,1,0,3,0,1,1,15,27,"(48, 16)"
-90,2,1,1,1,1,0,1,1,27,15,"(75, 26)"
-91,2,1,1,0,1,0,1,1,15,27,"(63, 42)"
-92,2,1,1,0,2,0,1,1,27,15,"(74, 10)"
-93,1,0,0,0,8,4,0,0,37,79,"(1, 4)"
-94,2,1,1,0,1,0,1,1,15,27,"(74, 75)"
-95,1,1,0,0,5,0,1,1,18,93,"(66, 83)"
-96,2,1,1,1,3,0,1,1,27,15,"(43, 49)"
-97,1,1,0,0,1,0,1,1,93,18,"(10, 39)"
-98,1,0,0,0,1,0,1,0,18,93,"(37, 36)"
-99,1,1,0,0,1,0,1,1,93,18,"(20, 34)"
-100,2,1,0,0,1,0,1,1,37,79,"(99, 83)"
