--- MAINCODE/results/agent_data_17.csv
@@ -1,71 +1,0 @@
-Number, Tribe, Condition, Visibility Range, Metabolism, Intelligent, Covered Distance, Reproduction Counter, Consume Counter, Sickness Counter, Parent A, Parent B, Position
-1,1,2,1,1,True,16,0,16,0,,,"(74, 81)"
-2,2,1,1,1,True,18,1,18,0,,,"(6, 7)"
-3,3,1,2,1,False,59,0,9,6,,,"(79, 14)"
-4,1,1,1,2,True,31,0,31,0,,,"(20, 56)"
-5,1,2,2,2,True,31,0,31,0,,,"(33, 0)"
-6,2,1,2,1,False,59,0,9,6,,,"(63, 51)"
-7,2,3,1,1,False,17,0,17,0,,,"(24, 59)"
-8,1,1,2,2,False,72,0,13,6,,,"(0, 49)"
-9,1,1,3,3,False,89,0,10,8,,,"(49, 5)"
-10,3,3,3,3,True,50,1,41,0,,,"(7, 8)"
-11,3,1,1,2,False,65,0,15,6,,,"(44, 12)"
-12,3,1,2,3,True,48,4,35,0,,,"(3, 6)"
-13,1,3,2,1,True,16,0,16,0,,,"(34, 59)"
-14,2,3,3,1,True,18,5,16,0,,,"(1, 4)"
-15,2,3,3,2,True,29,0,28,0,,,"(7, 9)"
-16,1,3,1,3,True,48,0,48,0,,,"(36, 2)"
-17,1,3,2,3,True,48,0,44,0,,,"(8, 5)"
-18,2,3,3,3,True,48,0,48,0,,,"(0, 29)"
-19,2,1,3,3,False,81,0,14,7,,,"(18, 21)"
-20,2,3,3,2,True,29,1,29,0,,,"(1, 23)"
-21,3,1,3,3,False,83,0,13,7,,,"(0, 48)"
-22,2,3,3,2,False,34,9,27,0,,,"(9, 4)"
-23,1,3,2,2,False,30,0,30,0,,,"(9, 3)"
-24,3,2,3,2,True,35,3,31,0,,,"(1, 21)"
-25,2,2,2,3,True,49,1,37,0,,,"(10, 4)"
-26,2,2,3,3,True,44,2,44,0,,,"(0, 30)"
-27,2,1,2,2,False,80,0,10,8,,,"(48, 50)"
-28,1,2,1,1,False,19,0,19,0,,,"(51, 63)"
-29,3,3,1,2,False,33,0,33,0,,,"(4, 15)"
-30,1,1,3,3,True,44,0,44,0,,,"(4, 14)"
-31,2,1,2,1,True,30,3,16,0,,,"(0, 4)"
-32,3,1,3,2,False,29,0,29,0,,,"(4, 16)"
-33,3,2,2,1,False,21,4,16,0,,,"(4, 12)"
-34,1,1,3,2,False,67,0,14,6,,,"(45, 31)"
-35,3,1,3,2,False,79,0,9,8,,,"(49, 35)"
-36,2,2,3,1,True,17,0,17,0,,,"(5, 6)"
-37,3,1,2,3,False,88,0,11,8,,,"(63, 72)"
-38,1,2,3,3,True,41,0,40,0,,,"(8, 8)"
-39,2,2,3,3,True,46,0,46,0,,,"(2, 18)"
-40,1,2,3,3,True,43,1,43,0,,,"(0, 35)"
-41,3,2,3,1,True,28,2,16,0,,,"(1, 0)"
-42,2,3,1,1,True,17,1,16,0,,,"(10, 2)"
-43,3,1,1,2,False,64,0,12,6,,,"(38, 37)"
-44,1,2,3,2,True,30,0,30,0,,,"(7, 10)"
-45,2,1,3,2,True,29,0,29,0,,,"(11, 1)"
-46,3,1,2,1,False,18,0,18,0,,,"(27, 50)"
-47,3,2,1,1,True,17,0,17,0,,,"(45, 41)"
-48,1,3,1,3,True,47,1,42,0,,,"(3, 6)"
-49,1,3,2,2,False,73,0,13,6,,,"(47, 46)"
-50,3,2,1,3,True,40,0,40,0,,,"(0, 56)"
-51,2,1,1,1,0,8,0,8,0,25,41,"(0, 78)"
-52,2,1,1,0,0,6,0,6,0,22,14,"(24, 5)"
-53,2,1,1,1,0,5,0,5,0,20,26,"(24, 6)"
-54,2,1,1,0,0,6,0,6,0,22,14,"(29, 77)"
-55,2,1,1,0,0,5,0,5,0,14,22,"(59, 42)"
-56,2,1,1,0,0,4,0,4,0,22,14,"(10, 3)"
-57,2,1,1,0,0,4,0,4,0,14,22,"(22, 7)"
-58,1,1,1,1,0,3,0,3,0,40,26,"(53, 53)"
-59,3,1,1,1,0,2,1,2,0,10,22,"(6, 5)"
-60,3,1,0,0,0,5,0,3,0,42,41,"(37, 67)"
-61,2,0,1,0,0,2,0,2,0,22,31,"(9, 16)"
-62,2,1,1,0,0,3,0,3,0,31,22,"(53, 10)"
-63,3,1,1,0,0,3,0,3,0,33,24,"(75, 20)"
-64,3,0,1,1,0,2,0,2,0,12,24,"(69, 14)"
-65,2,1,1,0,0,2,0,2,0,22,31,"(29, 65)"
-66,3,1,1,0,0,3,0,3,0,33,24,"(66, 47)"
-67,2,0,0,0,0,5,0,1,0,2,59,"(11, 67)"
-68,3,0,1,1,0,1,0,1,0,12,33,"(59, 12)"
-69,3,0,1,1,0,2,0,2,0,33,12,"(51, 54)"
-70,3,1,0,1,0,0,0,0,0,48,12,"(47, 14)"
