--- MAINCODE/results/agent_data_18.csv
@@ -1,54 +1,0 @@
-Number, Tribe, Condition, Visibility Range, Metabolism, Intelligent, Covered Distance, Reproduction Counter, Consume Counter, Sickness Counter, Parent A, Parent B, Position
-1,2,1,3,1,False,56,0,8,5,,,"(74, 0)"
-2,2,1,2,2,False,68,0,14,6,,,"(17, 22)"
-3,1,1,1,3,False,84,0,14,8,,,"(34, 1)"
-4,3,1,1,2,False,77,0,9,7,,,"(47, 13)"
-5,3,1,1,1,False,59,0,9,6,,,"(47, 78)"
-6,3,1,2,2,False,73,0,11,7,,,"(53, 21)"
-7,2,1,1,3,True,41,0,41,0,,,"(53, 47)"
-8,2,2,3,3,True,41,0,41,0,,,"(2, 13)"
-9,1,2,2,2,False,63,0,13,5,,,"(26, 0)"
-10,3,2,1,2,False,30,0,30,0,,,"(4, 52)"
-11,1,2,3,3,True,43,2,43,0,,,"(3, 8)"
-12,3,3,2,2,False,72,2,12,6,,,"(2, 8)"
-13,2,1,1,3,True,43,0,43,0,,,"(34, 13)"
-14,3,2,2,1,True,16,0,16,0,,,"(26, 13)"
-15,2,1,1,2,False,79,0,11,7,,,"(49, 84)"
-16,3,2,3,2,True,30,0,29,0,,,"(6, 1)"
-17,3,1,2,3,True,45,0,45,0,,,"(4, 5)"
-18,2,2,2,2,False,28,0,28,0,,,"(5, 47)"
-19,2,3,1,2,True,31,0,31,0,,,"(7, 6)"
-20,2,1,1,3,False,42,0,42,0,,,"(29, 57)"
-21,2,1,3,1,True,17,0,17,0,,,"(22, 25)"
-22,2,3,2,3,False,47,0,43,0,,,"(5, 4)"
-23,3,3,3,2,True,29,0,29,0,,,"(1, 28)"
-24,1,1,3,2,True,29,0,29,0,,,"(2, 14)"
-25,3,1,2,2,True,30,0,30,0,,,"(23, 1)"
-26,1,1,3,2,False,73,0,10,7,,,"(1, 6)"
-27,2,1,2,3,False,86,1,13,8,,,"(0, 23)"
-28,2,2,3,1,True,17,0,17,0,,,"(0, 40)"
-29,1,3,3,1,True,18,0,18,0,,,"(0, 38)"
-30,1,3,1,3,True,42,0,42,0,,,"(25, 8)"
-31,1,1,3,1,False,18,1,18,0,,,"(0, 18)"
-32,2,1,2,2,True,30,0,30,0,,,"(0, 20)"
-33,2,2,1,3,True,42,0,42,0,,,"(22, 1)"
-34,3,3,3,1,False,18,0,18,0,,,"(38, 0)"
-35,2,3,2,2,False,28,0,28,0,,,"(33, 1)"
-36,2,1,2,1,False,60,0,7,6,,,"(78, 69)"
-37,1,1,3,3,False,88,0,9,8,,,"(34, 0)"
-38,1,1,1,2,False,81,0,11,8,,,"(19, 1)"
-39,1,3,1,3,False,48,0,48,0,,,"(38, 24)"
-40,3,1,2,1,False,58,0,8,6,,,"(36, 80)"
-41,3,1,2,2,True,33,0,33,0,,,"(2, 12)"
-42,2,1,1,2,False,64,0,12,6,,,"(44, 11)"
-43,1,2,2,3,True,44,0,44,0,,,"(1, 17)"
-44,2,3,1,1,True,18,0,18,0,,,"(10, 1)"
-45,2,3,1,2,False,31,0,31,0,,,"(6, 2)"
-46,1,1,2,3,True,51,0,51,0,,,"(1, 15)"
-47,3,1,2,3,True,43,0,43,0,,,"(0, 21)"
-48,3,1,2,1,True,20,0,20,0,,,"(29, 2)"
-49,2,1,3,1,True,21,0,21,0,,,"(4, 6)"
-50,2,3,1,3,True,44,0,44,0,,,"(27, 51)"
-51,2,1,1,1,False,6,0,2,1,27,31,"(27, 61)"
-52,3,0,1,1,True,1,0,1,0,11,12,"(31, 19)"
-53,3,0,1,1,True,1,0,1,0,11,12,"(22, 86)"
