--- MAINCODE/results/agent_data_19.csv
@@ -1,67 +1,0 @@
-Number, Tribe, Condition, Visibility Range, Metabolism, Intelligent, Covered Distance, Reproduction Counter, Consume Counter, Sickness Counter, Parent A, Parent B, Position
-1,1,1,2,3,False,80,0,13,7,,,"(17, 51)"
-2,3,1,3,3,False,85,0,14,8,,,"(13, 18)"
-3,3,3,2,2,True,33,0,33,0,,,"(24, 11)"
-4,2,1,2,2,True,31,0,31,0,,,"(13, 0)"
-5,3,3,2,1,False,18,0,18,0,,,"(37, 57)"
-6,3,2,1,3,False,42,0,42,0,,,"(19, 1)"
-7,3,1,1,1,True,17,0,17,0,,,"(72, 3)"
-8,3,1,3,1,True,19,0,19,0,,,"(4, 1)"
-9,1,3,1,3,False,40,0,40,0,,,"(39, 60)"
-10,2,2,3,1,True,20,2,20,0,,,"(3, 6)"
-11,1,3,1,1,False,18,0,18,0,,,"(21, 8)"
-12,3,1,1,3,False,79,0,14,7,,,"(17, 65)"
-13,3,3,1,2,False,79,0,9,7,,,"(39, 0)"
-14,1,1,3,1,False,19,0,19,0,,,"(2, 4)"
-15,3,1,1,1,False,59,0,9,5,,,"(36, 51)"
-16,1,2,2,1,True,19,0,19,0,,,"(53, 0)"
-17,1,2,1,2,True,27,0,27,0,,,"(1, 26)"
-18,3,3,2,3,False,45,0,45,0,,,"(8, 0)"
-19,2,1,3,2,False,77,1,9,7,,,"(2, 8)"
-20,1,2,3,1,True,16,10,16,0,,,"(4, 5)"
-21,3,1,2,2,True,29,0,29,0,,,"(1, 50)"
-22,2,2,2,1,True,21,0,21,0,,,"(1, 45)"
-23,2,3,2,2,False,31,0,31,0,,,"(28, 0)"
-24,1,3,1,3,True,47,0,47,0,,,"(31, 2)"
-25,1,3,2,2,True,28,0,28,0,,,"(21, 40)"
-26,2,3,1,1,False,16,0,16,0,,,"(0, 25)"
-27,1,1,3,2,True,33,2,32,0,,,"(5, 0)"
-28,2,1,3,3,False,88,0,10,8,,,"(2, 9)"
-29,2,1,2,3,False,79,0,13,7,,,"(0, 73)"
-30,3,2,3,2,True,30,2,30,0,,,"(0, 11)"
-31,2,1,2,1,False,52,0,8,5,,,"(60, 16)"
-32,1,1,2,1,False,58,0,7,6,,,"(55, 16)"
-33,3,3,3,3,False,43,2,43,0,,,"(0, 12)"
-34,1,1,3,3,False,83,0,13,8,,,"(38, 35)"
-35,2,2,2,2,False,30,0,29,0,,,"(1, 10)"
-36,2,2,1,3,True,44,0,44,0,,,"(33, 35)"
-37,1,3,3,2,True,26,0,26,0,,,"(0, 16)"
-38,3,3,1,3,False,83,0,13,7,,,"(85, 54)"
-39,1,1,3,2,True,28,0,28,0,,,"(1, 42)"
-40,3,1,3,3,False,81,0,11,7,,,"(15, 33)"
-41,1,1,2,3,False,86,0,12,8,,,"(48, 31)"
-42,1,1,3,3,False,85,1,15,7,,,"(4, 3)"
-43,1,2,3,1,False,16,0,16,0,,,"(29, 1)"
-44,3,1,2,1,True,17,0,17,0,,,"(0, 69)"
-45,1,1,1,1,True,16,0,16,0,,,"(77, 0)"
-46,3,2,2,1,True,17,0,17,0,,,"(56, 59)"
-47,2,2,3,1,False,17,0,17,0,,,"(39, 8)"
-48,2,1,1,2,True,31,0,31,0,,,"(0, 37)"
-49,1,2,1,3,True,46,2,35,0,,,"(4, 4)"
-50,1,1,2,3,False,76,0,16,7,,,"(50, 65)"
-51,2,1,1,0,False,32,0,4,3,10,30,"(62, 85)"
-52,2,1,1,0,False,31,0,3,3,30,10,"(59, 53)"
-53,1,0,0,1,False,13,10,1,1,49,27,"(3, 2)"
-54,1,0,1,1,True,3,0,3,0,27,49,"(60, 49)"
-55,3,1,1,1,True,4,0,4,0,42,33,"(88, 80)"
-56,2,0,1,1,True,3,0,3,0,19,33,"(88, 65)"
-57,1,0,0,0,True,3,0,1,0,20,53,"(29, 94)"
-58,1,1,0,0,False,1,0,1,1,53,20,"(81, 86)"
-59,1,1,1,0,False,2,0,1,1,20,53,"(4, 26)"
-60,1,1,0,0,False,1,0,1,1,53,20,"(63, 53)"
-61,1,1,0,0,False,1,0,1,1,20,53,"(84, 13)"
-62,1,0,0,0,True,3,0,0,0,53,20,"(5, 28)"
-63,1,0,0,0,True,1,0,1,0,20,53,"(67, 59)"
-64,1,1,0,0,False,1,0,1,1,53,20,"(8, 44)"
-65,1,0,0,0,True,1,0,1,0,20,53,"(76, 7)"
-66,1,0,0,0,False,1,0,1,0,53,20,"(59, 33)"
