--- MAINCODE/results/agent_data_2.csv
@@ -1,11 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Position
-1,1,2,2,0,10,,,"(0, 87)"
-2,1,3,1,0,10,,,"(1, 12)"
-3,1,2,2,0,10,,,"(35, 11)"
-4,2,1,3,0,10,,,"(28, 42)"
-5,2,3,2,0,10,,,"(16, 50)"
-6,2,3,1,0,10,,,"(14, 29)"
-7,1,2,3,0,10,,,"(16, 64)"
-8,2,3,3,0,10,,,"(57, 63)"
-9,3,2,1,0,10,,,"(9, 90)"
-10,1,1,2,0,10,,,"(55, 36)"
