--- MAINCODE/results/agent_data_3.csv
@@ -1,92 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Position
-1,3,3,1,2,79,,,"(2, 16)"
-2,2,1,1,0,95,,,"(22, 5)"
-3,2,1,1,0,88,,,"(6, 45)"
-4,2,1,1,0,79,,,"(2, 37)"
-5,2,1,3,2,70,,,"(5, 10)"
-6,1,2,1,2,89,,,"(3, 14)"
-7,2,3,1,5,65,,,"(8, 7)"
-8,1,3,2,2,88,,,"(9, 6)"
-9,1,2,2,4,59,,,"(5, 3)"
-10,3,3,2,0,100,,,"(1, 79)"
-11,3,3,1,0,87,,,"(3, 32)"
-12,3,3,3,2,83,,,"(12, 24)"
-13,3,1,3,3,84,,,"(5, 6)"
-14,1,3,2,3,70,,,"(19, 14)"
-15,1,2,3,0,75,,,"(10, 30)"
-16,2,3,3,0,97,,,"(10, 33)"
-17,1,1,1,0,88,,,"(11, 11)"
-18,2,3,1,0,84,,,"(4, 10)"
-19,1,2,3,4,79,,,"(3, 6)"
-20,1,2,2,3,57,,,"(6, 10)"
-21,2,3,2,3,75,,,"(14, 19)"
-22,2,1,3,3,67,,,"(8, 10)"
-23,2,3,3,5,54,,,"(0, 6)"
-24,3,2,3,3,69,,,"(13, 11)"
-25,2,1,2,7,38,,,"(1, 0)"
-26,1,3,3,0,100,,,"(10, 35)"
-27,3,3,2,1,93,,,"(0, 49)"
-28,2,1,1,0,93,,,"(14, 20)"
-29,3,2,2,2,53,,,"(18, 11)"
-30,1,2,3,0,99,,,"(1, 80)"
-31,2,2,2,1,68,,,"(1, 12)"
-32,3,2,2,3,73,,,"(11, 4)"
-33,1,2,3,0,90,,,"(6, 32)"
-34,2,2,1,0,95,,,"(7, 55)"
-35,2,3,1,1,79,,,"(21, 9)"
-36,1,2,1,1,69,,,"(13, 23)"
-37,3,2,1,1,83,,,"(24, 0)"
-38,3,3,3,2,95,,,"(10, 36)"
-39,2,3,3,0,93,,,"(0, 21)"
-40,3,2,2,1,91,,,"(4, 33)"
-41,1,3,1,0,94,,,"(3, 69)"
-42,3,1,1,2,52,,,"(6, 18)"
-43,2,2,3,2,70,,,"(10, 12)"
-44,1,3,1,0,90,,,"(4, 25)"
-45,3,3,3,0,100,,,"(0, 83)"
-46,1,2,3,0,82,,,"(10, 37)"
-47,2,3,2,3,78,,,"(13, 22)"
-48,3,3,3,2,74,,,"(3, 25)"
-49,2,2,2,1,69,,,"(2, 25)"
-50,3,3,2,0,94,,,"(14, 16)"
-51,2,1,1,0,65,21,48,"(23, 3)"
-52,3,0,0,0,6,1,42,"(82, 13)"
-53,2,1,0,0,22,37,7,"(48, 17)"
-54,1,1,1,1,40,8,14,"(15, 8)"
-55,3,1,1,3,32,24,12,"(6, 17)"
-56,1,1,1,0,54,29,20,"(22, 45)"
-57,1,1,1,1,46,9,14,"(14, 6)"
-58,3,1,1,0,50,38,27,"(35, 2)"
-59,2,0,1,0,49,9,25,"(34, 0)"
-60,3,0,1,0,40,42,24,"(5, 47)"
-61,2,0,1,0,36,25,43,"(2, 51)"
-62,3,0,1,0,35,29,22,"(25, 2)"
-63,3,0,1,0,19,13,32,"(8, 28)"
-64,1,1,1,0,23,9,23,"(8, 49)"
-65,3,1,1,0,15,32,14,"(8, 29)"
-66,1,1,1,0,22,20,19,"(61, 69)"
-67,1,1,1,0,21,31,8,"(2, 75)"
-68,3,0,1,0,20,5,24,"(44, 3)"
-69,3,1,1,0,20,13,1,"(31, 64)"
-70,1,1,1,0,20,20,19,"(28, 3)"
-71,3,0,0,0,2,21,55,"(90, 54)"
-72,2,1,0,0,5,7,47,"(23, 99)"
-73,1,0,0,1,0,36,55,"(15, 7)"
-74,2,1,0,0,6,7,47,"(93, 42)"
-75,3,1,0,0,5,32,35,"(49, 52)"
-76,2,0,1,0,13,54,22,"(70, 6)"
-77,2,0,0,0,1,7,25,"(69, 57)"
-78,3,1,1,0,6,40,38,"(24, 31)"
-79,3,0,1,0,5,9,13,"(1, 82)"
-80,2,0,1,0,5,23,25,"(73, 38)"
-81,2,0,1,0,5,25,23,"(28, 58)"
-82,1,1,1,0,3,19,6,"(36, 54)"
-83,2,0,1,0,3,23,25,"(77, 75)"
-84,2,1,0,0,2,55,21,"(82, 91)"
-85,3,1,1,0,2,12,47,"(52, 12)"
-86,1,1,0,0,1,19,6,"(84, 32)"
-87,2,1,1,0,2,25,23,"(82, 40)"
-88,2,1,1,0,2,49,48,"(32, 34)"
-89,1,0,0,0,0,57,73,"(20, 2)"
-90,2,0,1,0,0,5,7,"(52, 99)"
-91,2,0,1,0,0,22,43,"(93, 98)"
