--- MAINCODE/results/agent_data_4.csv
@@ -1,120 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Parent A,Parent B,Position
-1,2,1,1,0,88,,,"(15, 15)"
-2,2,3,3,1,81,,,"(11, 28)"
-3,1,2,1,0,94,,,"(13, 3)"
-4,1,3,1,3,54,,,"(3, 8)"
-5,3,3,3,7,66,,,"(13, 19)"
-6,2,3,1,4,61,,,"(10, 34)"
-7,1,3,2,2,87,,,"(10, 29)"
-8,1,2,3,1,79,,,"(3, 11)"
-9,2,1,1,0,81,,,"(13, 2)"
-10,3,2,3,3,76,,,"(10, 30)"
-11,1,2,2,0,100,,,"(5, 53)"
-12,3,3,1,3,76,,,"(3, 6)"
-13,2,1,2,2,63,,,"(1, 10)"
-14,3,2,1,5,48,,,"(0, 0)"
-15,3,2,1,3,77,,,"(9, 13)"
-16,3,2,3,8,68,,,"(0, 2)"
-17,1,1,1,0,89,,,"(5, 48)"
-18,3,3,3,0,97,,,"(10, 33)"
-19,2,3,1,2,88,,,"(6, 41)"
-20,1,3,1,2,89,,,"(9, 25)"
-21,1,2,3,2,95,,,"(10, 35)"
-22,1,3,3,3,76,,,"(11, 30)"
-23,3,2,2,0,100,,,"(5, 57)"
-24,3,3,3,7,73,,,"(3, 8)"
-25,3,2,2,6,67,,,"(0, 2)"
-26,2,2,2,2,84,,,"(7, 21)"
-27,3,1,1,9,60,,,"(1, 0)"
-28,3,3,1,3,71,,,"(12, 29)"
-29,3,2,3,3,81,,,"(4, 27)"
-30,2,1,2,3,68,,,"(14, 12)"
-31,3,2,2,0,100,,,"(6, 45)"
-32,3,2,1,0,97,,,"(14, 14)"
-33,3,1,3,2,68,,,"(2, 7)"
-34,3,3,1,0,89,,,"(6, 50)"
-35,3,1,3,3,87,,,"(16, 12)"
-36,3,2,3,0,98,,,"(5, 54)"
-37,2,2,3,4,70,,,"(4, 5)"
-38,1,2,2,1,68,,,"(0, 24)"
-39,1,1,3,3,63,,,"(16, 11)"
-40,1,1,2,2,55,,,"(1, 17)"
-41,2,1,2,2,95,,,"(1, 43)"
-42,1,2,2,5,51,,,"(8, 7)"
-43,3,1,3,4,66,,,"(2, 2)"
-44,2,1,1,1,79,,,"(1, 29)"
-45,1,1,2,1,63,,,"(10, 16)"
-46,3,3,1,4,83,,,"(0, 32)"
-47,1,2,3,0,79,,,"(16, 16)"
-48,3,3,1,0,87,,,"(11, 39)"
-49,2,2,3,3,67,,,"(17, 17)"
-50,3,2,2,2,87,,,"(2, 27)"
-51,3,1,0,2,19,29,15,"(23, 0)"
-52,2,1,0,3,22,6,15,"(7, 40)"
-53,3,1,1,2,73,10,25,"(22, 2)"
-54,2,0,0,0,5,52,50,"(38, 44)"
-55,3,1,1,0,66,10,5,"(12, 18)"
-56,2,0,0,0,3,52,41,"(89, 16)"
-57,2,1,0,0,27,6,37,"(51, 54)"
-58,3,1,1,0,61,35,28,"(17, 21)"
-59,2,0,0,0,4,52,41,"(57, 71)"
-60,3,1,1,3,25,24,5,"(21, 0)"
-61,3,1,1,0,51,24,5,"(20, 5)"
-62,3,1,1,0,50,24,5,"(39, 3)"
-63,3,1,1,1,30,5,24,"(7, 13)"
-64,3,0,1,0,47,27,33,"(40, 0)"
-65,3,0,1,0,46,40,29,"(29, 1)"
-66,3,1,1,0,44,42,5,"(55, 30)"
-67,1,1,0,0,19,7,20,"(34, 83)"
-68,2,0,1,0,27,49,30,"(16, 8)"
-69,1,1,1,0,36,42,49,"(51, 0)"
-70,2,1,1,0,35,37,2,"(10, 38)"
-71,1,1,1,0,29,6,22,"(13, 28)"
-72,2,0,0,0,2,27,30,"(69, 84)"
-73,2,0,1,0,28,13,37,"(2, 65)"
-74,3,0,0,0,2,51,60,"(94, 52)"
-75,2,0,1,0,31,26,40,"(56, 42)"
-76,3,0,0,0,2,51,60,"(89, 35)"
-77,2,1,0,0,13,6,22,"(78, 82)"
-78,3,1,0,2,11,53,12,"(67, 11)"
-79,1,0,1,0,26,42,39,"(16, 25)"
-80,2,1,0,0,9,46,19,"(4, 88)"
-81,1,1,0,0,11,38,46,"(54, 7)"
-82,3,1,1,1,16,10,8,"(6, 41)"
-83,1,0,1,0,23,45,21,"(23, 44)"
-84,2,0,0,2,2,27,13,"(69, 13)"
-85,3,1,0,0,8,28,46,"(23, 29)"
-86,2,0,1,0,22,30,49,"(64, 1)"
-87,2,0,1,0,21,37,39,"(49, 55)"
-88,3,0,1,0,21,43,35,"(58, 27)"
-89,3,0,1,0,20,16,35,"(1, 77)"
-90,3,0,1,0,20,16,43,"(54, 4)"
-91,3,1,0,0,9,42,14,"(55, 99)"
-92,3,1,0,0,7,25,14,"(32, 29)"
-93,3,1,0,0,6,25,14,"(93, 81)"
-94,3,0,0,0,1,16,27,"(88, 77)"
-95,3,1,0,1,0,12,16,"(9, 2)"
-96,3,1,0,0,6,15,24,"(69, 68)"
-97,3,0,1,0,11,27,16,"(50, 18)"
-98,3,0,0,0,1,25,27,"(99, 53)"
-99,3,0,0,0,1,95,27,"(98, 32)"
-100,2,1,0,0,0,28,26,"(8, 1)"
-101,3,0,1,0,6,43,25,"(65, 53)"
-102,3,0,0,0,1,60,53,"(10, 42)"
-103,1,1,1,0,5,21,7,"(25, 12)"
-104,1,1,0,0,3,42,4,"(70, 8)"
-105,1,1,0,0,2,4,39,"(75, 74)"
-106,3,0,0,0,1,16,27,"(47, 4)"
-107,1,1,1,0,4,22,20,"(40, 70)"
-108,3,1,0,0,2,27,12,"(84, 63)"
-109,2,0,0,0,1,44,46,"(54, 45)"
-110,3,1,1,0,3,50,29,"(74, 39)"
-111,3,0,0,0,1,78,84,"(48, 96)"
-112,2,0,0,0,1,84,78,"(90, 92)"
-113,3,1,1,0,2,63,5,"(30, 82)"
-114,3,0,1,0,1,14,43,"(63, 8)"
-115,3,1,1,0,0,14,16,"(55, 82)"
-116,3,1,0,0,0,24,4,"(27, 43)"
-117,3,0,1,0,0,24,33,"(47, 73)"
-118,3,1,1,0,0,25,16,"(65, 45)"
-119,3,0,0,0,0,82,19,"(98, 4)"
