--- MAINCODE/results/agent_data_5.csv
@@ -1,1 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
