--- MAINCODE/results/agent_data_6.csv
@@ -1,51 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
-1,3,3,3,0,0,0,,,"(30, 73)"
-2,2,1,2,0,0,0,,,"(80, 83)"
-3,2,1,1,0,0,0,,,"(37, 88)"
-4,2,3,1,0,0,0,,,"(1, 65)"
-5,1,3,2,0,0,0,,,"(48, 45)"
-6,1,2,2,0,0,0,,,"(39, 42)"
-7,3,3,2,0,0,0,,,"(61, 34)"
-8,1,2,2,0,0,0,,,"(21, 48)"
-9,3,2,2,0,0,0,,,"(74, 32)"
-10,2,3,1,0,0,0,,,"(36, 68)"
-11,1,1,2,0,0,0,,,"(31, 4)"
-12,3,3,3,0,0,0,,,"(0, 10)"
-13,3,2,3,0,0,0,,,"(56, 56)"
-14,2,1,2,0,0,0,,,"(86, 30)"
-15,2,1,2,0,0,0,,,"(34, 38)"
-16,2,2,1,0,0,0,,,"(84, 27)"
-17,1,2,2,0,0,0,,,"(59, 94)"
-18,2,2,2,0,0,0,,,"(10, 63)"
-19,1,2,1,0,0,0,,,"(54, 29)"
-20,1,3,2,0,0,0,,,"(36, 86)"
-21,2,3,2,0,0,0,,,"(74, 62)"
-22,1,1,3,0,0,0,,,"(28, 87)"
-23,1,1,1,0,0,0,,,"(80, 3)"
-24,3,1,2,0,0,0,,,"(25, 87)"
-25,1,2,2,0,0,0,,,"(52, 37)"
-26,1,1,1,0,0,0,,,"(13, 8)"
-27,1,2,1,0,0,0,,,"(76, 86)"
-28,3,1,1,0,0,0,,,"(95, 81)"
-29,2,3,2,0,0,0,,,"(67, 27)"
-30,2,3,3,0,0,0,,,"(82, 62)"
-31,2,3,3,0,0,0,,,"(2, 23)"
-32,2,1,2,0,0,0,,,"(3, 17)"
-33,1,3,1,0,0,0,,,"(38, 54)"
-34,3,2,1,0,0,0,,,"(72, 70)"
-35,2,3,3,0,0,0,,,"(53, 99)"
-36,2,3,3,0,0,0,,,"(11, 87)"
-37,3,1,1,0,0,0,,,"(72, 92)"
-38,2,1,2,0,0,0,,,"(74, 49)"
-39,1,3,3,0,0,0,,,"(61, 20)"
-40,1,1,3,0,0,0,,,"(96, 62)"
-41,2,2,3,0,0,0,,,"(30, 85)"
-42,3,2,2,0,0,0,,,"(7, 18)"
-43,3,3,3,0,0,0,,,"(70, 61)"
-44,2,3,2,0,0,0,,,"(36, 3)"
-45,1,1,1,0,0,0,,,"(51, 35)"
-46,3,1,1,0,0,0,,,"(0, 84)"
-47,1,2,1,0,0,0,,,"(91, 14)"
-48,3,2,1,0,0,0,,,"(21, 72)"
-49,1,2,1,0,0,0,,,"(78, 89)"
-50,2,2,2,0,0,0,,,"(76, 51)"
