--- MAINCODE/results/agent_data_7.csv
@@ -1,51 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
-1,3,2,1,0,0,0,,,"(53, 76)"
-2,2,1,2,0,0,0,,,"(55, 12)"
-3,2,2,3,0,0,0,,,"(22, 40)"
-4,3,1,3,0,0,0,,,"(96, 19)"
-5,3,2,3,0,0,0,,,"(4, 96)"
-6,3,2,1,0,0,0,,,"(74, 89)"
-7,3,2,1,0,0,0,,,"(21, 35)"
-8,2,3,3,0,0,0,,,"(60, 45)"
-9,2,1,1,0,0,0,,,"(52, 80)"
-10,3,1,1,0,0,0,,,"(94, 25)"
-11,3,3,1,0,0,0,,,"(23, 22)"
-12,3,2,3,0,0,0,,,"(24, 50)"
-13,3,2,2,0,0,0,,,"(42, 97)"
-14,2,1,1,0,0,0,,,"(72, 69)"
-15,2,3,2,0,0,0,,,"(81, 3)"
-16,3,1,2,0,0,0,,,"(22, 39)"
-17,3,2,2,0,0,0,,,"(28, 43)"
-18,2,2,3,0,0,0,,,"(64, 37)"
-19,1,3,3,0,0,0,,,"(0, 26)"
-20,1,1,2,0,0,0,,,"(63, 60)"
-21,1,1,1,0,0,0,,,"(34, 54)"
-22,1,1,3,0,0,0,,,"(7, 33)"
-23,2,2,2,0,0,0,,,"(32, 89)"
-24,1,2,1,0,0,0,,,"(65, 56)"
-25,3,1,3,0,0,0,,,"(49, 6)"
-26,1,3,2,0,0,0,,,"(49, 18)"
-27,3,3,1,0,0,0,,,"(55, 27)"
-28,3,3,3,0,0,0,,,"(19, 65)"
-29,2,1,2,0,0,0,,,"(9, 67)"
-30,3,2,3,0,0,0,,,"(79, 71)"
-31,2,1,3,0,0,0,,,"(41, 50)"
-32,1,1,2,0,0,0,,,"(53, 89)"
-33,2,1,3,0,0,0,,,"(63, 94)"
-34,3,3,1,0,0,0,,,"(72, 77)"
-35,3,3,3,0,0,0,,,"(22, 23)"
-36,2,2,3,0,0,0,,,"(20, 84)"
-37,3,3,1,0,0,0,,,"(32, 31)"
-38,3,2,1,0,0,0,,,"(24, 83)"
-39,1,2,1,0,0,0,,,"(34, 63)"
-40,3,3,1,0,0,0,,,"(90, 70)"
-41,2,3,3,0,0,0,,,"(72, 22)"
-42,3,3,2,0,0,0,,,"(7, 53)"
-43,1,3,3,0,0,0,,,"(1, 15)"
-44,1,2,3,0,0,0,,,"(27, 43)"
-45,3,1,3,0,0,0,,,"(5, 65)"
-46,1,1,2,0,0,0,,,"(28, 73)"
-47,1,1,2,0,0,0,,,"(31, 96)"
-48,1,2,1,0,0,0,,,"(80, 30)"
-49,2,2,3,0,0,0,,,"(79, 26)"
-50,3,3,1,0,0,0,,,"(3, 67)"
