--- MAINCODE/results/agent_data_8.csv
@@ -1,51 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
-1,3,2,1,0,0,0,,,"(6, 20)"
-2,1,1,3,0,0,0,,,"(67, 24)"
-3,2,3,2,0,0,0,,,"(37, 84)"
-4,3,2,3,0,0,0,,,"(5, 38)"
-5,2,1,3,0,0,0,,,"(41, 71)"
-6,2,1,2,0,0,0,,,"(86, 60)"
-7,2,2,2,0,0,0,,,"(47, 20)"
-8,3,3,3,0,0,0,,,"(22, 71)"
-9,3,1,2,0,0,0,,,"(88, 51)"
-10,3,2,3,0,0,0,,,"(64, 17)"
-11,3,2,1,0,0,0,,,"(21, 24)"
-12,3,2,1,0,0,0,,,"(78, 50)"
-13,3,3,1,0,0,0,,,"(6, 5)"
-14,2,2,2,0,0,0,,,"(48, 61)"
-15,1,2,1,0,0,0,,,"(93, 95)"
-16,2,3,3,0,0,0,,,"(86, 28)"
-17,3,1,2,0,0,0,,,"(31, 72)"
-18,1,1,3,0,0,0,,,"(55, 63)"
-19,3,1,2,0,0,0,,,"(3, 33)"
-20,3,3,2,0,0,0,,,"(20, 88)"
-21,2,1,2,0,0,0,,,"(78, 28)"
-22,3,2,1,0,0,0,,,"(99, 30)"
-23,1,3,3,0,0,0,,,"(17, 98)"
-24,1,1,3,0,0,0,,,"(85, 45)"
-25,2,3,2,0,0,0,,,"(71, 7)"
-26,1,1,2,0,0,0,,,"(70, 90)"
-27,2,1,1,0,0,0,,,"(97, 2)"
-28,1,1,3,0,0,0,,,"(85, 66)"
-29,2,2,3,0,0,0,,,"(3, 48)"
-30,3,1,2,0,0,0,,,"(32, 54)"
-31,3,2,2,0,0,0,,,"(76, 24)"
-32,1,3,1,0,0,0,,,"(77, 82)"
-33,1,3,1,0,0,0,,,"(97, 11)"
-34,2,1,3,0,0,0,,,"(27, 94)"
-35,3,2,3,0,0,0,,,"(80, 34)"
-36,1,1,3,0,0,0,,,"(26, 68)"
-37,2,2,2,0,0,0,,,"(30, 98)"
-38,1,1,1,0,0,0,,,"(91, 34)"
-39,3,1,2,0,0,0,,,"(20, 11)"
-40,1,3,3,0,0,0,,,"(59, 58)"
-41,1,1,3,0,0,0,,,"(71, 41)"
-42,1,2,1,0,0,0,,,"(43, 80)"
-43,3,1,3,0,0,0,,,"(43, 57)"
-44,1,3,2,0,0,0,,,"(31, 74)"
-45,3,1,2,0,0,0,,,"(97, 12)"
-46,1,2,1,0,0,0,,,"(92, 53)"
-47,3,1,2,0,0,0,,,"(78, 45)"
-48,1,3,3,0,0,0,,,"(85, 36)"
-49,2,2,2,0,0,0,,,"(41, 75)"
-50,1,1,1,0,0,0,,,"(9, 98)"
