--- MAINCODE/results/agent_data_9.csv
@@ -1,57 +1,0 @@
-Number,Tribe,Condition,Visibility Range,Reproduction Counter,Consume Counter,Sickness Counter,Parent A,Parent B,Position
-1,2,3,1,0,0,0,,,"(7, 34)"
-2,2,3,1,0,0,0,,,"(75, 97)"
-3,2,1,1,0,0,0,,,"(32, 51)"
-4,3,2,1,0,0,0,,,"(57, 81)"
-5,2,2,2,0,0,0,,,"(15, 48)"
-6,1,2,1,0,0,0,,,"(17, 36)"
-7,3,3,2,0,0,0,,,"(4, 88)"
-8,2,3,2,0,0,0,,,"(17, 56)"
-9,1,2,3,0,0,0,,,"(88, 4)"
-10,3,2,1,0,0,0,,,"(82, 16)"
-11,3,3,2,0,0,0,,,"(47, 79)"
-12,2,3,3,0,0,0,,,"(25, 99)"
-13,3,1,1,0,0,0,,,"(40, 37)"
-14,2,1,1,0,0,0,,,"(44, 14)"
-15,3,2,2,6,0,0,,,"(97, 54)"
-16,1,3,3,0,0,0,,,"(3, 59)"
-17,2,1,1,0,0,0,,,"(18, 50)"
-18,2,1,1,6,0,0,,,"(97, 54)"
-19,1,2,2,0,0,0,,,"(19, 24)"
-20,3,1,3,0,0,0,,,"(50, 68)"
-21,1,2,3,0,0,0,,,"(85, 89)"
-22,1,2,1,0,0,0,,,"(61, 42)"
-23,3,1,2,0,0,0,,,"(32, 14)"
-24,3,1,2,0,0,0,,,"(99, 49)"
-25,3,2,1,0,0,0,,,"(92, 61)"
-26,2,1,1,0,0,0,,,"(81, 31)"
-27,3,1,1,0,0,0,,,"(77, 84)"
-28,1,1,2,0,0,0,,,"(86, 30)"
-29,1,1,2,0,0,0,,,"(99, 26)"
-30,1,1,2,0,0,0,,,"(8, 20)"
-31,2,2,3,0,0,0,,,"(61, 39)"
-32,2,2,3,0,0,0,,,"(21, 64)"
-33,3,2,1,0,0,0,,,"(1, 19)"
-34,1,1,3,0,0,0,,,"(60, 89)"
-35,2,3,2,0,0,0,,,"(10, 0)"
-36,2,2,1,0,0,0,,,"(34, 68)"
-37,3,3,3,0,0,0,,,"(31, 10)"
-38,1,2,1,0,0,0,,,"(69, 24)"
-39,3,3,1,0,0,0,,,"(9, 85)"
-40,3,2,3,0,0,0,,,"(57, 16)"
-41,3,1,3,0,0,0,,,"(52, 27)"
-42,2,3,1,0,0,0,,,"(8, 27)"
-43,1,1,1,0,0,0,,,"(14, 37)"
-44,2,3,1,0,0,0,,,"(28, 19)"
-45,1,3,3,0,0,0,,,"(62, 97)"
-46,2,2,1,0,0,0,,,"(84, 87)"
-47,2,2,1,0,0,0,,,"(56, 35)"
-48,1,1,2,0,0,0,,,"(76, 37)"
-49,1,1,1,0,0,0,,,"(79, 43)"
-50,3,2,1,0,0,0,,,"(18, 4)"
-51,2,0,0,0,0,0,15,18,"(39, 87)"
-52,3,0,0,0,0,0,18,15,"(47, 11)"
-53,2,0,0,0,0,0,15,18,"(71, 2)"
-54,3,0,0,0,0,0,18,15,"(16, 3)"
-55,3,0,0,0,0,0,18,15,"(17, 69)"
-56,3,0,0,0,0,0,15,18,"(43, 33)"
