--- pytest-Genetic_Algorithm/MAINCODE/results/agent_data_17.csv
@@ -1,68 +1,0 @@
-Number, Tribe, Condition, Visibility Range, Metabolism, Covered Distance, Reproduction Counter, Consume Counter, Sickness Counter, Parent A, Parent B, Position
-1,1,3,2,1,16,0,16,0,,,"(0, 51)"
-2,1,2,3,1,60,0,10,5,,,"(12, 33)"
-3,1,1,1,1,16,0,16,0,,,"(39, 1)"
-4,2,1,2,2,72,0,12,7,,,"(31, 75)"
-5,1,1,1,1,60,0,10,6,,,"(21, 14)"
-6,3,2,3,1,59,0,9,5,,,"(25, 35)"
-7,3,2,1,2,29,0,29,0,,,"(38, 59)"
-8,3,2,3,3,43,0,43,0,,,"(2, 4)"
-9,1,2,3,1,18,0,18,0,,,"(1, 42)"
-10,2,3,3,2,80,7,10,7,,,"(33, 0)"
-11,1,1,3,1,62,0,10,6,,,"(69, 59)"
-12,2,1,2,3,48,6,48,0,,,"(3, 3)"
-13,1,1,2,2,80,0,10,7,,,"(3, 39)"
-14,1,1,1,3,84,0,13,8,,,"(0, 84)"
-15,3,1,1,1,59,0,9,6,,,"(38, 51)"
-16,2,1,2,1,60,0,9,6,,,"(4, 1)"
-17,2,3,1,3,42,0,42,0,,,"(4, 2)"
-18,2,1,1,2,73,0,13,7,,,"(0, 72)"
-19,2,3,2,1,17,0,17,0,,,"(0, 59)"
-20,1,1,3,3,87,0,11,8,,,"(44, 0)"
-21,3,3,2,1,18,0,18,0,,,"(1, 11)"
-22,2,1,1,2,74,0,12,7,,,"(8, 7)"
-23,2,1,3,1,58,0,9,5,,,"(69, 46)"
-24,3,2,1,3,46,0,46,0,,,"(0, 40)"
-25,3,1,1,1,43,0,13,4,,,"(60, 22)"
-26,3,1,2,1,56,0,9,5,,,"(0, 2)"
-27,1,2,2,2,30,0,30,0,,,"(32, 21)"
-28,3,1,3,3,85,1,15,7,,,"(2, 9)"
-29,1,1,1,1,60,0,7,6,,,"(18, 14)"
-30,3,2,2,1,18,0,18,0,,,"(63, 41)"
-31,3,1,1,3,84,0,14,8,,,"(66, 64)"
-32,1,2,3,1,18,0,18,0,,,"(2, 6)"
-33,3,1,2,1,56,0,7,5,,,"(47, 62)"
-34,2,1,2,1,44,0,10,4,,,"(38, 52)"
-35,1,1,2,1,60,0,8,6,,,"(29, 76)"
-36,2,1,2,3,88,0,10,8,,,"(0, 56)"
-37,1,1,1,3,89,0,13,8,,,"(45, 58)"
-38,2,1,3,1,51,0,11,5,,,"(20, 0)"
-39,2,1,2,3,76,7,15,7,,,"(36, 1)"
-40,3,2,1,1,68,0,8,6,,,"(57, 74)"
-41,2,1,2,3,41,4,41,0,,,"(0, 19)"
-42,3,1,1,2,67,0,14,6,,,"(0, 89)"
-43,3,2,2,3,45,0,45,0,,,"(1, 14)"
-44,2,1,1,3,87,0,9,8,,,"(75, 76)"
-45,2,1,2,1,50,0,10,5,,,"(0, 47)"
-46,2,1,2,1,60,0,10,6,,,"(68, 44)"
-47,3,1,2,2,75,0,10,7,,,"(12, 43)"
-48,3,1,2,1,59,0,9,5,,,"(0, 77)"
-49,3,3,3,3,85,9,15,7,,,"(1, 15)"
-50,1,3,2,1,21,0,21,0,,,"(19, 16)"
-51,3,0,1,1,13,0,3,1,12,49,"(36, 62)"
-52,2,1,1,1,12,0,2,2,49,12,"(44, 16)"
-53,2,1,1,1,17,0,2,2,49,12,"(42, 30)"
-54,3,0,1,1,12,0,2,1,12,49,"(58, 58)"
-55,3,1,1,1,12,0,2,2,49,12,"(25, 71)"
-56,2,1,1,1,17,0,2,2,10,39,"(93, 24)"
-57,2,1,1,1,7,0,3,1,39,10,"(68, 8)"
-58,2,0,1,1,12,0,2,1,10,39,"(56, 11)"
-59,2,1,1,1,4,0,4,1,39,10,"(3, 5)"
-60,2,1,1,1,3,0,3,1,10,39,"(92, 75)"
-61,2,1,1,1,4,0,4,1,39,10,"(17, 66)"
-62,2,0,1,1,12,0,2,1,10,39,"(94, 22)"
-63,2,1,1,1,6,0,2,1,28,12,"(5, 38)"
-64,2,0,1,1,1,0,1,0,41,49,"(25, 21)"
-65,2,1,1,1,1,0,1,1,41,49,"(83, 84)"
-66,3,1,1,1,1,0,1,1,41,49,"(28, 61)"
-67,3,1,1,1,1,0,1,1,49,41,"(55, 59)"
