+++ results/agent_data_0.csv
@@ -1,0 +1,13 @@
+Number, Tribe, Condition, Visibility Range, Metabolism, Intelligent, Aggressive, Covered Distance, Reproduction Counter, Consume Counter, Sickness Counter, Parent A, Parent B, Position, Expelled
+1,3,1,1,1,False,True,2,0,0,0,,,"(2, 1)",0
+2,3,3,1,1,False,True,2,0,0,0,,,"(9, 6)",0
+3,1,1,2,2,True,False,2,0,0,0,,,"(9, 6)",0
+4,2,3,1,1,False,True,2,1,0,0,,,"(0, 3)",0
+5,3,2,1,2,False,True,2,0,0,0,,,"(7, 7)",0
+6,3,3,3,3,True,False,2,0,0,0,,,"(3, 6)",0
+7,3,3,3,1,False,True,2,1,0,0,,,"(3, 1)",0
+8,2,3,2,3,False,True,2,0,0,0,,,"(3, 0)",0
+9,3,1,3,3,False,True,2,1,0,0,,,"(4, 4)",0
+10,1,1,2,2,False,True,2,0,0,0,,,"(8, 5)",0
+11,3,1,1,1,True,False,1,1,0,0,7,9,"(0, 3)",0
+12,2,0,0,0,False,True,0,0,0,0,11,4,"(1, 5)",0
